requirements.txt
 Django==2.2.5
 django_taggit==0.23
 djangorestframework==3.9.1
 django-taggit-serializer==0.1.7
 django-cors-headers==3.1.0
 django-debug-toolbar==2.0